[Analytics Backend / DataFusion] Move partialAggOutputSchema to ExchangeSinkProvider
`FragmentConvertor` is a Calcite RelNode → backend-bytes converter; asking it
for the schema of already-converted bytes inverted that contract. Move the
method onto `ExchangeSinkProvider` (the SPI for coordinator-side sinks),
which is what `LocalStageScheduler` already pulls from the stage.

Drops `Stage.fragmentConvertor` field/accessors and the wire-up in
`FragmentConversionDriver` — that plumbing existed only to thread the
convertor reference for a non-conversion concern.
